flask-api/requirements.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
librosa==0.8.0
llvmlite==0.34.0
MarkupSafe==1.1.1
numba==0.51.2
numpy==1.19.4
packaging==20.4
pooch==1.2.0
pycparser==2.20
pyparsing==2.4.7
python-firebase==1.2
requests==2.24.0
resampy==0.2.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
urllib3==1.25.11
Werkzeug==1.0.1
